Deployment/requirements.txt
scikit-learn
pandas
matplotlib
joblib
transformers
seaborn
streamlit
tensorflow
keras
numpy
Pillow
torchvision